+++ Sources/EudiWalletKit/Services/OpenId4VciService.swift
@@ -77,21 +77,25 @@ public class OpenId4VCIService: NSObject, ASWebAuthenticationPresentationContext
 	///   - uriOffer: Uri of the offer (from a QR or a deep link)
 	///   - format: format of the exchanged data
 	/// - Returns: The data of the document
-	public func resolveOfferDocTypes(uriOffer: String, format: DataFormat = .cbor) async throws -> OfferedIssueModel {
+	public func resolveOfferDocTypes(uriOffer: String, format: DataFormat = .cbor) async throws -> OfferedIssuanceModel {
 		let result = await CredentialOfferRequestResolver(fetcher: Fetcher(session: urlSession), credentialIssuerMetadataResolver: CredentialIssuerMetadataResolver(fetcher: Fetcher(session: urlSession)), authorizationServerMetadataResolver: AuthorizationServerMetadataResolver(oidcFetcher: Fetcher(session: urlSession), oauthFetcher: Fetcher(session: urlSession))).resolve(source: try .init(urlString: uriOffer))
 		switch result {
 		case .success(let offer):
-			let code: Grants.PreAuthorizedCode? = switch offer.grants {	case .preAuthorizedCode(let preAuthorizedCode):	preAuthorizedCode; case .both(_, let preAuthorizedCode):	preAuthorizedCode; case .authorizationCode(_), .none: nil	}
+			let code: Grants.PreAuthorizedCode? = switch offer.grants {	case .preAuthorizedCode(let preAuthorizedCode):	preAuthorizedCode; case .both(_, let preAuthorizedCode): preAuthorizedCode; case .authorizationCode(_), .none: nil	}
 			Self.metadataCache[uriOffer] = offer
 			let credentialInfo = try getCredentialIdentifiers(credentialsSupported: offer.credentialIssuerMetadata.credentialsSupported.filter { offer.credentialConfigurationIdentifiers.contains($0.key) }, format: format)
-			return OfferedIssueModel(issuerName: offer.credentialIssuerIdentifier.url.absoluteString, docModels: credentialInfo.map(\.offered), txCodeSpec:  code?.txCode)
+			return OfferedIssuanceModel(issuerName: offer.credentialIssuerIdentifier.url.absoluteString, docModels: credentialInfo.map(\.offered), txCodeSpec:  code?.txCode)
 		case .failure(let error):
 			throw WalletError(description: "Unable to resolve credential offer: \(error.localizedDescription)")
 		}
 	}
 	
-	public func getIssuer(offer: CredentialOffer) throws -> Issuer {
+	func getIssuer(offer: CredentialOffer) throws -> Issuer {
 		try Issuer(authorizationServerMetadata: offer.authorizationServerMetadata, issuerMetadata: offer.credentialIssuerMetadata, config: config, parPoster: Poster(session: urlSession), tokenPoster: Poster(session: urlSession), requesterPoster: Poster(session: urlSession), deferredRequesterPoster: Poster(session: urlSession), notificationPoster: Poster(session: urlSession))
+	}
+	
+	func getIssuerForDeferred(data: DeferredIssuanceModel) throws -> Issuer {
+		try Issuer(authorizationServerMetadata: .oauth(AuthorizationServerMetadata(issuer: nil, authorizationEndpoint: nil, tokenEndpoint: nil, introspectionEndpoint: nil, jwksURI: nil, grantTypesSupported: nil, responseTypesSupported: nil, requestObjectSigningAlgValuesSupported: nil, requestObjectEncryptionAlgValuesSupported: nil, requestObjectEncryptionEncValuesSupported: nil, responseModesSupported: nil, registrationEndpoint: nil, tokenEndpointAuthMethodsSupported: nil, tokenEndpointAuthSigningAlgValuesSupported: nil, introspectionEndpointAuthMethodsSupported: nil, introspectionEndpointAuthSigningAlgValuesSupported: nil, authorizationSigningAlgValuesSupported: nil, authorizationEncryptionAlgValuesSupported: nil, authorizationEncryptionEncValuesSupported: nil, scopesSupported: nil, requestParameterSupported: nil, requestURIParameterSupported: nil, requireRequestURIRegistration: nil, codeChallengeMethodsSupported: nil, tlsClientCertificateBoundAccessTokens: nil, dpopSigningAlgValuesSupported: nil, revocationEndpoint: nil, revocationEndpointAuthMethodsSupported: nil, revocationEndpointAuthSigningAlgValuesSupported: nil, deviceAuthorizationEndpoint: nil, backchannelTokenDeliveryModesSupported: nil, backchannelAuthenticationEndpoint: nil, backchannelAuthenticationRequestSigningAlgValuesSupported: nil, requirePushedAuthorizationRequests: nil, pushedAuthorizationRequestEndpoint: nil, mtlsEndpointAliases: nil, authorizationResponseIssParameterSupported: nil)), issuerMetadata: CredentialIssuerMetadata(credentialIssuerIdentifier: CredentialIssuerId(""), authorizationServers: [], credentialEndpoint: data.deferredCredentialEndpoint, batchCredentialEndpoint: nil, deferredCredentialEndpoint: data.deferredCredentialEndpoint, notificationEndpoint: nil, credentialConfigurationsSupported: [:], signedMetadata: nil, display: nil), config: config, parPoster: Poster(session: urlSession), tokenPoster: Poster(session: urlSession), requesterPoster: Poster(session: urlSession), deferredRequesterPoster: Poster(session: urlSession), notificationPoster: Poster(session: urlSession))
 	}
 	
 	func issueDocumentsByOfferUrl(offerUri: String, docTypes: [OfferedDocModel], txCodeValue: String?, format: DataFormat, useSecureEnclave: Bool = true, claimSet: ClaimSet? = nil) async throws -> [IssuanceOutcome] {
@@ -105,7 +109,7 @@ public class OpenId4VCIService: NSObject, ASWebAuthenticationPresentationContext
 		let preAuthorizedCode: String? = code?.preAuthorizedCode
 		let issuer = try getIssuer(offer: offer)
 		if preAuthorizedCode != nil && txCodeSpec != nil && txCodeValue == nil { throw WalletError(description: "A transaction code is required for this offer") }
-		let authorized = if let preAuthorizedCode, let txCodeValue, let authCode = try? IssuanceAuthorization(preAuthorizationCode: preAuthorizedCode, txCode: txCodeSpec) { try await issuer.authorizeWithPreAuthorizationCode(credentialOffer: offer, authorizationCode: authCode, clientId: config.clientId, transactionCode: txCodeValue).get() } else { try await authorizeRequestWithAuthCodeUseCase(issuer: issuer, offer: offer) }
+		let authorized = if let preAuthorizedCode, let authCode = try? IssuanceAuthorization(preAuthorizationCode: preAuthorizedCode, txCode: txCodeSpec) { try await issuer.authorizeWithPreAuthorizationCode(credentialOffer: offer, authorizationCode: authCode, clientId: config.clientId, transactionCode: txCodeValue).get() } else { try await authorizeRequestWithAuthCodeUseCase(issuer: issuer, offer: offer) }
 		let data = await credentialInfo.asyncCompactMap {
 			do {
 				logger.info("Starting issuing with identifer \($0.identifier.value) and scope \($0.scope)")
@@ -234,7 +238,7 @@ public class OpenId4VCIService: NSObject, ASWebAuthenticationPresentationContext
 	
 	private func noProofRequiredSubmissionUseCase(issuer: Issuer, noProofRequiredState: AuthorizedRequest, credentialConfigurationIdentifier: CredentialConfigurationIdentifier, claimSet: ClaimSet? = nil) async throws -> IssuanceOutcome {
 		switch noProofRequiredState {
-		case .noProofRequired(let accessToken, let refreshToken, let credentialIdentifiers):
+		case .noProofRequired(let accessToken, let refreshToken, _):
 			let payload: IssuanceRequestPayload = .configurationBased(credentialConfigurationIdentifier: credentialConfigurationIdentifier,	claimSet: claimSet)
 			let responseEncryptionSpecProvider = { Issuer.createResponseEncryptionSpec($0) }
 			let requestOutcome = try await issuer.requestSingle(noProofRequest: noProofRequiredState, requestPayload: payload, responseEncryptionSpecProvider: responseEncryptionSpecProvider)
@@ -246,7 +250,7 @@ public class OpenId4VCIService: NSObject, ASWebAuthenticationPresentationContext
 						switch result {
 						case .deferred(let transactionId):
 							//return try await deferredCredentialUseCase(issuer: issuer, authorized: noProofRequiredState, transactionId: transactionId)
-							let deferredModel = DeferredIssuanceModel(credentialIssuerUrl: credentialIssuerURL, accessToken: accessToken, refreshToken: refreshToken, credentialIdentifiers: credentialIdentifiers, transactionId: transactionId)
+							let deferredModel = await DeferredIssuanceModel(deferredCredentialEndpoint: issuer.issuerMetadata.deferredCredentialEndpoint!, accessToken: accessToken, refreshToken: refreshToken, transactionId: transactionId)
 							return .deferred(deferredModel)
 						case .issued(_, let credential, _):
 							guard let data = Data(base64URLEncoded: credential) else { throw WalletError(description: "Invalid credential")	}
@@ -281,7 +285,7 @@ public class OpenId4VCIService: NSObject, ASWebAuthenticationPresentationContext
 					switch result {
 					case .deferred(let transactionId):
 						//return try await deferredCredentialUseCase(issuer: issuer, authorized: authorized, transactionId: transactionId)
-						let deferredModel = DeferredIssuanceModel(credentialIssuerUrl: credentialIssuerURL, accessToken: accessToken, refreshToken: refreshToken, credentialIdentifiers: credentialIdentifiers, transactionId: transactionId)
+						let deferredModel = await DeferredIssuanceModel(deferredCredentialEndpoint: issuer.issuerMetadata.deferredCredentialEndpoint!, accessToken: accessToken, refreshToken: refreshToken, transactionId: transactionId)
 						return .deferred(deferredModel)
 					case .issued(_, let credential, _):
 						guard let data = Data(base64URLEncoded: credential) else { throw WalletError(description: "Invalid credential")	}
@@ -311,10 +315,10 @@ public class OpenId4VCIService: NSObject, ASWebAuthenticationPresentationContext
 			case .issuancePending(let transactionId):
 				logger.info("Credential not ready yet. Try after \(transactionId.interval ?? 0)")
 				let deferredModel = switch authorized {
-				case .noProofRequired(let accessToken, let refreshToken, let credentialIdentifiers):
-					DeferredIssuanceModel(credentialIssuerUrl: credentialIssuerURL, accessToken: accessToken, refreshToken: refreshToken, credentialIdentifiers: credentialIdentifiers, transactionId: transactionId)
-				case .proofRequired(let accessToken, let refreshToken, _, let credentialIdentifiers):
-					DeferredIssuanceModel(credentialIssuerUrl: credentialIssuerURL, accessToken: accessToken, refreshToken: refreshToken, credentialIdentifiers: credentialIdentifiers, transactionId: transactionId)
+				case .noProofRequired(let accessToken, let refreshToken, _):
+					await DeferredIssuanceModel(deferredCredentialEndpoint: issuer.issuerMetadata.deferredCredentialEndpoint!, accessToken: accessToken, refreshToken: refreshToken, transactionId: transactionId)
+				case .proofRequired(let accessToken, let refreshToken, _, _):
+					await DeferredIssuanceModel(deferredCredentialEndpoint: issuer.issuerMetadata.deferredCredentialEndpoint!, accessToken: accessToken, refreshToken: refreshToken, transactionId: transactionId)
 				}
 				return .deferred(deferredModel)
 			case .errored(_, let errorDescription):